Move display_anim into _core
