+++ ecell4/plotting/_core.py
@@ -1,3 +1,7 @@
+import os
+import os.path
+from tempfile import NamedTemporaryFile
+
 def get_range_of_world(world, scale=1.0):
     edge_lengths = world.edge_lengths() * scale
     max_length = max(tuple(edge_lengths))
@@ -68,6 +72,46 @@ def get_range_of_trajectories(data, plot_range=None):
         raise ValueError(
             'plot_range must be list, tuple or dict. [{}] was given.'.format(
                 repr(plot_range)))
+
+def anim_to_html(anim, filename=None, fps=6, crf=10, bitrate='1M'):
+    VIDEO_TAG = """<video controls>
+     <source src="data:video/x-webm;base64,{0}" type="video/webm">
+     Your browser does not support the video tag.
+    </video>"""
+    import base64
+
+    if not hasattr(anim, '_encoded_video'):
+        if filename is None:
+            f = NamedTemporaryFile(suffix='.webm', delete=False)
+            filename = f.name
+            f.close()
+            # anim.save(filename, fps=fps, extra_args=['-vcodec', 'libvpx'])
+            anim.save(filename, fps=fps, codec='libvpx', extra_args=['-auto-alt-ref', '0', '-crf', str(crf), '-b:v', bitrate])
+            # anim.save(filename, writer='mencoder', fps=fps, extra_args=['-lavcopts', 'vcodec=libvpx'])
+            video = open(filename, "rb").read()
+            os.remove(filename)
+            # with NamedTemporaryFile(suffix='.webm') as f:
+            #     anim.save(f.name, fps=fps, extra_args=['-vcodec', 'libvpx'])
+            #     video = open(f.name, "rb").read()
+        else:
+            with open(filename, 'w') as f:
+                anim.save(f.name, fps=fps, extra_args=['-vcodec', 'libvpx'])
+                video = open(f.name, "rb").read()
+        # anim._encoded_video = video.encode("base64")
+        anim._encoded_video = base64.encodestring(video).decode('utf-8')
+    return VIDEO_TAG.format(anim._encoded_video)
+
+def display_anim(ani, output=None, fps=6, crf=10, bitrate='1M'):
+    if output is None:
+        from IPython.display import display, HTML
+        display(HTML(anim_to_html(ani, output, fps=fps, crf=crf, bitrate=bitrate)))
+    elif os.path.splitext(output.lower())[1] == '.webm':
+        ani.save(output, fps=fps, codec='libvpx', extra_args=['-auto-alt-ref', '0', '-crf', str(crf), '-b:v', bitrate])
+    elif os.path.splitext(output.lower())[1] == '.mp4':
+        ani.save(output, fps=fps, codec='mpeg4', extra_args=['-crf', str(crf), '-b:v', bitrate])
+    else:
+        raise ValueError(
+            "An output filename is only accepted with extension '.webm' or 'mp4'.")
 
 from ..util.model_parser import Visitor, dispatch, RATELAW_RESERVED_FUNCTIONS, RATELAW_RESERVED_CONSTANTS
 import ecell4_base
